New Update : Added Csv Helper. Not done with the BulkInsertOperation Function

+++ MP-MMS.WPF/Views/Windows/AddPart.xaml.cs
@@ -53,7 +53,7 @@ namespace MP_MMS.WPF.Views.Windows
             if (cBoxLocation.SelectedItem != null) { part.LocationId = ((Location)cBoxLocation.SelectedItem).Id; }
 
 
-            part.UnitCost = (double)txtCostPerUnit.Value;
+            part.UnitCost = (decimal)txtCostPerUnit.Value;
             part.Quantity = (int)txtQuantity.Value;
             part.DateAdded = Convert.ToDateTime(txtDateAdded.Value);
